+++ app/src/main/AndroidManifest.xml
@@ -82,7 +82,7 @@
         <activity
             android:name=".ui.GroupManagerActivity"
             android:label="@string/title_activity_manage_groups" />
-        <activity android:name=".AssignIconsActivity"
+        <activity android:name=".ui.AssignIconsActivity"
             android:label="@string/title_activity_assign_icons"/>
         <activity
             android:name=".ui.PanicResponderActivity"
