+++ app/src/main/res/layout/activity_assign_icons.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?attr/background"
-    tools:context="com.beemdevelopment.aegis.AssignIconsActivity">
+    tools:context="com.beemdevelopment.aegis.ui.AssignIconsActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
