+++ app/src/main/res/menu/menu_assign_icons.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context="com.beemdevelopment.aegis.ui.AssignIconsActivity">
+    <item
+        android:id="@+id/action_save"
+        app:showAsAction="always"
+        android:title="@string/save"/>
+</menu>
